+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [0.8.4] - 2023-10-12
+
+### Added
+- exposed the tryAdd method of request headers through request info.
+
+### Changed
+- Defaults the content type parameter in `setStreamContent` to `application/octet-stream`
+
 ## [0.8.3] - 2023-10-11
 
 ### Added